refactor: code formats
