+++ src/common/Domain/Repositories/GetInfoRepo.ts
@@ -1,16 +1,15 @@
 import axios, { AxiosError, AxiosResponse } from 'axios';
 import { BASEURL, GET_INFO } from '../../../lib/constants/urls';
+import { logError } from '../../LogError';
 
 export const getInfo = (completion: (data: any) => void) => {
   axios
     .get(BASEURL + GET_INFO)
     .then((response: AxiosResponse) => {
-      console.log(response.status);
-
       const { data } = response;
       completion(data);
     })
     .catch((error: AxiosError) => {
-      console.log(error.message);
+      logError(error)
     });
 };
