+++ src/common/Domain/Repositories/LogOutRepo.ts
@@ -1,16 +1,15 @@
 import axios, { AxiosError, AxiosResponse } from 'axios';
 import { BASEURL, LOGOUT } from '../../../lib/constants/urls';
+import { logError } from '../../LogError';
 
 export const logout = (completion: (data: any) => void) => {
   axios
     .get(BASEURL + LOGOUT)
     .then((response: AxiosResponse) => {
-      console.log(response.status);
-
       const { data } = response;
       completion(data);
     })
     .catch((error: AxiosError) => {
-      console.log(error.message);
+      logError(error);
     });
 };
