+++ src/common/Domain/Repositories/SubjectRepo.ts
@@ -1,16 +1,15 @@
 import axios, { AxiosError, AxiosResponse } from 'axios';
 import { BASEURL, GET_SUBJECTS } from '../../../lib/constants/urls';
 import { SubjectData } from '../Entities/SubjectEntity';
+import { logError } from '../../LogError';
 
 export const fetchSubject = (completion: (data: SubjectData[]) => void) => {
   axios
     .get(BASEURL + GET_SUBJECTS)
     .then((response: AxiosResponse) => {
-      console.log(response);
       completion(response.data.data);
     })
     .catch((error: AxiosError) => {
-      /* eslint-disable no-console */
-      console.log(error);
+      logError(error);
     });
 };
