+++ src/common/LogError.ts
@@ -1,0 +1,5 @@
+import { AxiosError } from "axios"
+
+export const logError = (error: AxiosError) => {
+  console.log(error);
+}
